Add timestamp lock

+++ Assets/Scripts/Notes/NoteData.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [System.Serializable]
 public class NoteData
 {
-    public float timeToHit;  // when to hit
+    public double timeToHit;  // when to hit
     // no lane or groupIndex here anymore
 }
